\docType{data}
\name{CGH_annot}
\alias{CGH_annot}
\title{Gene expression annotations}
\format{one instance, 1 row per probe}
\source{
  IGR, Villejuif, France
}
\description{
  This data set contains annotations on the probes used for
  comparative genomics hybridizatin
}
\examples{
\dontrun{
library(gliomaData)
#read data
data(CGH_annot)
}
}
\author{
  V Frouin, 2012-07-31
}
\keyword{datasets}

