\docType{data}
\name{clinic}
\alias{clinic}
\title{Clinic from the patient data file}
\format{one instance, 1 row per patient}
\source{
  IGR, Villejuif, France
}
\description{
  This data
}
\examples{
\dontrun{
library(gliomaData)
#read data
data(clinic)
}
}
\author{
  V Frouin, 2012-07-31
}
\keyword{datasets}

