% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GE_annot.R
\docType{data}
\name{GE_annot}
\alias{GE_annot}
\title{Gene expression annotations}
\format{
one instance, 1 row per probe
}
\source{
IGR, Villejuif, France
}
\usage{
GE_annot
}
\description{
This data set contains annotations on the probes used for
gene expression
}
\examples{
#read data
data(GE_annot)
head(GE_annot)
}
\author{
V Frouin, 2012-07-31
}
