\docType{data}
\name{histo}
\alias{histo}
\title{histo from patient data file}
\format{one instance, 1 row per patient}
\source{
  IGR, Villejuif, France
}
\description{
  This data set contains information on the histology
}
\examples{
\dontrun{
library(gliomaData)
#read data
data(histo)
}
}
\author{
  V Frouin, 2012-07-31
}
\keyword{datasets}

