% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/img.R
\docType{data}
\name{img}
\alias{img}
\title{Scoring from the neuroimaging of the patient}
\format{
one instance, 1 row per patient
}
\source{
IGR, Villejuif, France
}
\usage{
img
}
\description{
This data set contains information on the patient file
}
\examples{
data(img)
head(img)
}
\author{
V Frouin, 2012-07-31
}
