% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv.R
\docType{data}
\name{surv}
\alias{surv}
\title{Survival information from on the patient}
\format{one instance, 1 row per patient}
\source{
IGR, Villejuif, France
}
\usage{
surv
}
\description{
This data set contains survival censored information
}
\examples{
data(surv)
head(surv)
}
\author{
V Frouin, 2012-07-31
}
