% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CGH_annot.R
\docType{data}
\name{CGH_annot}
\alias{CGH_annot}
\title{Gene expression annotations}
\format{one instance, 1 row per probe}
\source{
IGR, Villejuif, France
}
\usage{
CGH_annot
}
\description{
This data set contains annotations on the probes used for
comparative genomics hybridizatin
}
\examples{
data(CGH_annot)
head(CGH_annot)
}
\author{
V Frouin, 2012-07-31
}
