% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinic.R
\docType{data}
\name{clinic}
\alias{clinic}
\title{Clinic from the patient data file}
\format{
one instance, 1 row per patient
}
\source{
IGR, Villejuif, France
}
\usage{
clinic
}
\description{
This data
}
\examples{
data(clinic)
head(clinic)
}
\author{
V Frouin, 2012-07-31
}
