\docType{data}
\name{img}
\alias{img}
\title{Scoring from the neuroimaging of the patient}
\format{one instance, 1 row per patient}
\source{
  IGR, Villejuif, France
}
\description{
  This data set contains information on the patient file
}
\examples{
\dontrun{
library(gliomaData)
#read data
data(img)
}
}
\author{
  V Frouin, 2012-07-31
}
\keyword{datasets}

