\docType{data}
\name{surv}
\alias{surv}
\title{Survival information from on the patient}
\format{one instance, 1 row per patient}
\source{
  IGR, Villejuif, France
}
\description{
  This data set contains survival censored information
}
\examples{
\dontrun{
library(gliomaData)
#read data
data(surv)
}
}
\author{
  V Frouin, 2012-07-31
}
\keyword{datasets}

